Added Modernizr (custom build)
I added Modernizr, but I'm not sure we really need it. There is no need
to check for features because we are not supporting older browsers. If
someone is using an older browser that does not support CSS Gradients,
we can't imagine we'll implement polyfills for that user.

That being said, I check for CSS Gradients and Local Storage, Session
Storage, and Indexed DB. I figure the last three could be used for
allowing the user to store his patterns locally.

If you think I missed any necessary features feel free to add them. If
you agree we don't need Modernizr, let's dump it and save the page
speed.