Working on some bugs and trying to make strip movable
I almost merged ui and main js file. The reason of doing this was
easier development for me. In my IDE when an object is in the file it
can autocomplete for me. For example, now I can write ui. and IDE will
show me all ui properties which is very useful for me.

In end project we can separate files that is not a problem.